restructure for tests
