+++ README.md
@@ -1,0 +1,14 @@
+moved tests inside blackJack directory
+
+
+run code:
+
+from root directory `python blackJack/game.py`
+from blackJack directory, run game with `python game.py`  
+
+
+run tests:
+
+from blackJack directory, run tests with `python -m unittest test.test_blackjack`
+
+Do you need all of the int(17) calls? 17 is already an int. You don't need to convert it.
